pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[project]
name = "pyhaopenmotics"
description = "Asynchronous Python client for the OpenMotics API."
readme = "README.md"
keywords = [ "api", "async", "client", "homeassistant", "openmotics" ]
license = { text = "AGPLv3" }
maintainers = [ { name = "Wouter Coppens", email = "wouter.coppens@gmail.com" } ]
authors = [ { name = "Wouter Coppens", email = "wouter.coppens@gmail.com" } ]
requires-python = ">=3.14"
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: AsyncIO",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dynamic = [ "version" ]
dependencies = [
  "aiohttp>=3.11",
  "mashumaro>=3",
  "orjson>=3.10",
  "stamina>=25",
  "websockets>=15",
  "yarl>=1.18",
]
optional-dependencies.build = [ "hatch" ]
optional-dependencies.dev = [
  "authlib",
  "deptry",
  "hatch",
  "httpx",
  "pre-commit",
  "pyright",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-mock",
  "python-dotenv",
  "ruff",
  "tox-uv",
]
urls.Changelog = "https://github.com/rensongroup/pyhaopenmotics/releases"
urls.Issues = "https://github.com/rensongroup/pyhaopenmotics/issues"
urls.Repository = "https://github.com/rensongroup/pyhaopenmotics"

[tool.hatch]
version.path = "src/pyhaopenmotics/__version__.py"

[tool.ruff]
target-version = "py314"
line-length = 125  # Allow lines to be as long as 120.
fix = true
format.preview = true
lint.select = [
  # flake8-builtins
  "A",
  # flake8-bugbear
  "B",
  # flake8-comprehensions
  "C4",
  # mccabe
  "C90",
  # pycodestyle
  "E",
  # pyflakes
  "F",
  # isort
  "I",
  # pygrep-hooks
  "PGH",
  # ruff
  "RUF",
  # flake8-bandit
  "S",
  # flake8-simplify
  "SIM",
  # flake8-debugger
  "T10",
  # tryceratops
  "TRY",
  # pyupgrade
  "UP",
  "W",
  # flake8-2020
  "YTT",
]
lint.ignore = [
  # LineTooLong
  "E501",
  # DoNotAssignLambda
  "E731",
]
lint.per-file-ignores."tests/*" = [ "S101" ]

[tool.codespell]
write-changes = true

[tool.pytest]
ini_options.testpaths = [ "tests" ]
ini_options.python_files = [ "test_*.py" ]
ini_options.addopts = "-v --cov=pyhaopenmotics --cov-report=term-missing"
ini_options.asyncio_mode = "auto"

[tool.coverage]
run.branch = true
run.source = [ "src/pyhaopenmotics" ]
report.skip_empty = true

# [tool.coverage]
# run.plugins = ["covdefaults"]
# run.source = ["pyhaopenmotics"]
# report.fail_under = 25
# report.show_missing = true

[tool.pyright]
venvPath = "."
venv = ".venv"
include = [ "src" ]
exclude = [ "examples", "tests" ]
reportArgumentType = false
reportCallIssue = false
reportIncompatibleMethodOverride = false

[tool.bandit]
exclude_dirs = [ "tests" ]
# tests = ["B201", "B301"]
skips = [ "B101", "B601" ]
